enhancement(#4843): Removing uninstall at the end, API password and user as parameter
