+++ deployability/modules/testing/tests/helpers/generic.py
@@ -519,4 +519,27 @@ class GeneralComponentActions:
         Returns:
             bool: True/False
         """
-        return 'active' == Executor.execute_command(inventory_path, f'systemctl is-active {host_role}') 
+        return 'active' == Executor.execute_command(inventory_path, f'systemctl is-active {host_role}')
+
+
+
+class Waits:
+
+    @staticmethod
+    def dynamic_wait(expected_condition_func, cycles=10, waiting_time=10) -> None:
+        """
+        Waits the process during assigned cycles for the assigned seconds
+
+        Args:
+            expected_condition_func (lambda function): The function that returns True when the expected condition is met
+            cycles(int): Number of cycles
+            waiting_Time(int): Number of seconds per cycle
+
+        """
+        for _ in range(cycles):
+            if expected_condition_func():
+                break
+            else:
+                time.sleep(waiting_time)
+        else:
+            raise RuntimeError(f'Time out')
